refactor(GUI): generalize card (deck, faceup and placement) construction and rendering

+++ src/main/java/it/polimi/ingsw/am01/client/gui/controller/component/PlayablePositionController.java
@@ -5,7 +5,7 @@ import javafx.scene.shape.Rectangle;
 
 public class PlayablePositionController extends Rectangle implements ComponentController {
 
-    public PlayablePositionController(int x, int y) {
+    public PlayablePositionController(double x, double y) {
         loadComponent();
         setPosition(x, y);
     }
@@ -14,7 +14,7 @@ public class PlayablePositionController extends Rectangle implements ComponentCo
     private void initialize() {
     }
 
-    public void setPosition(int x, int y) {
+    public void setPosition(double x, double y) {
         this.setLayoutX(x);
         this.setLayoutY(y);
     }
